Evals/MultimodalRAG/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
langchain>=0.1.0
langchain-community>=0.0.10
pypdf2>=3.0.1
python-dotenv>=1.0.0
jinaai>=0.2.0
numpy>=1.24.3
python-multipart>=0.0.6
setuptools>=68.0.0
wheel>=0.40.0
